WIP: extract some common functionality
