+++ src/libertem_live/detectors/merlin/acquisition.py
@@ -1,33 +1,38 @@
 import logging
 from typing import Optional, NamedTuple
-from collections.abc import Generator, Iterator
+from collections.abc import Iterator
 import numpy as np
-from libertem.common import Shape, Slice
+from libertem.common import Shape
 from libertem.common.executor import (
-    TaskProtocol, WorkerQueue, TaskCommHandler, WorkerContext,
-    JobCancelledError,
+    WorkerContext,
 )
 from libertem.io.dataset.base import (
-    DataTile, DataSetMeta, BasePartition, Partition, DataSet, TilingScheme,
+    DataSetMeta, BasePartition, Partition, DataSet, TilingScheme,
 )
 from sparseconverter import ArrayBackend, NUMPY, CUDA
 
-from libertem_live.detectors.base.acquisition import AcquisitionMixin
+from libertem_live.detectors.base.acquisition import (
+    AcquisitionMixin, GenericCommHandler, GetFrames,
+)
 from libertem_live.hooks import Hooks, ReadyForDataEnv
-from .data import MerlinRawFrames, MerlinFrameStream, AcquisitionHeader
+from .data import AcquisitionHeader
 from .connection import MerlinDetectorConnection, MerlinPendingAcquisition
 from .controller import MerlinActiveController
 
 from opentelemetry import trace
-
+import libertem_qd_mpx
 
 tracer = trace.get_tracer(__name__)
 logger = logging.getLogger(__name__)
 
 
+class QdGetFrames(GetFrames):
+    CAM_CLIENT_CLS = libertem_qd_mpx.CamClient
+    FRAME_STACK_CLS = libertem_qd_mpx.QdFrameStack
+
+
 class AcqState(NamedTuple):
     acq_header: AcquisitionHeader
-    stream: MerlinFrameStream
 
 
 class MerlinAcquisition(AcquisitionMixin, DataSet):
@@ -139,25 +144,20 @@ class MerlinAcquisition(AcquisitionMixin, DataSet):
         if self._pending_aq is None:
             # active case, we manage the connection:
             self._conn.connect()
-            self._conn.maybe_drain()
             with tracer.start_as_current_span("MerlinAcquisition.trigger"):
                 self._hooks.on_ready_for_data(ReadyForDataEnv(aq=self))
-            acq_header, stream = self._conn.get_header_and_stream()
+
+            acq_header = self._conn.wait_for_acquisition()
+
             self._acq_state = AcqState(
                 acq_header=acq_header,
-                stream=stream,
             )
         else:
             # passive case, we don't manage the connection and we definitely
             # don't drain anything out of the socket!
             acq_header = self._pending_aq.header
-            stream = MerlinFrameStream.from_frame_header(
-                raw_socket=self._conn.get_data_socket(),
-                acquisition_header=acq_header
-            )
             self._acq_state = AcqState(
                 acq_header=acq_header,
-                stream=stream,
             )
 
     def end_acquisition(self):
@@ -203,51 +203,10 @@ class MerlinAcquisition(AcquisitionMixin, DataSet):
                 partition_slice=part_slice,
             )
 
-    def get_task_comm_handler(self) -> TaskCommHandler:
-        assert self._acq_state is not None
+    def get_task_comm_handler(self) -> "MerlinCommHandler":
         return MerlinCommHandler(
             conn=self._conn,
-            state=self._acq_state,
-            tiling_depth=24,  # FIXME!
         )
-
-
-def get_frames_from_queue(
-    queue: WorkerQueue,
-    tiling_scheme: TilingScheme,
-    sig_shape: tuple[int, ...],
-    dtype
-) -> Generator[tuple[np.ndarray, int], None, None]:
-    out = np.zeros((tiling_scheme.depth,) + sig_shape, dtype=dtype)
-    out_flat = out.reshape((tiling_scheme.depth, -1,))
-    while True:
-        with queue.get() as msg:
-            header, payload = msg
-            header_type = header["type"]
-            if header_type == "FRAME":
-                raw_frames = MerlinRawFrames(
-                    buffer=payload,
-                    start_idx=header['start_idx'],
-                    end_idx=header['end_idx'],
-                    first_frame_header=header['first_frame_header'],
-                )
-                with tracer.start_as_current_span("MerlinRawFrames.decode") as span:
-                    span.set_attributes({
-                        "libertem_live.header.start_idx": header["start_idx"],
-                        "libertem_live.header.end_idx": header["end_idx"],
-                        "libertem_live.buffer.len": len(payload),
-                    })
-                    raw_frames.decode(
-                        out_flat=out_flat,
-                    )
-                yield out[:raw_frames.num_frames], raw_frames.start_idx
-            elif header_type == "END_PARTITION":
-                # print(f"partition {partition} done")
-                return
-            else:
-                raise RuntimeError(
-                    f"invalid header type {header['type']}; FRAME or END_PARTITION expected"
-                )
 
 
 def _accum_partition(
@@ -268,69 +227,9 @@ def _accum_partition(
     return out
 
 
-class MerlinCommHandler(TaskCommHandler):
-    def __init__(
-        self,
-        conn: MerlinDetectorConnection,
-        state: AcqState,
-        tiling_depth: int
-    ):
-        self._conn = conn
-        self._acq_state = state
-        self._tiling_depth = tiling_depth
-
-    def handle_task(self, task: TaskProtocol, queue: WorkerQueue):
-        with tracer.start_as_current_span("MerlinCommHandler.handle_task") as span:
-            slice_ = task.get_partition().slice
-            start_idx = slice_.origin[0]
-            end_idx = slice_.origin[0] + slice_.shape[0]
-            span.set_attributes({
-                "libertem.partition.start_idx": start_idx,
-                "libertem.partition.end_idx": end_idx,
-            })
-            frame_chunk_size = min(32, self._tiling_depth)
-            frames_read = 0
-            num_frames_in_partition = end_idx - start_idx
-            stream = self._acq_state.stream
-            while frames_read < num_frames_in_partition:
-                next_read_size = min(
-                    num_frames_in_partition - frames_read,
-                    frame_chunk_size,
-                )
-                # FIXME: can't properly reuse this buffer currently...
-                input_buffer = stream.get_input_buffer(next_read_size)
-                res = stream.read_multi_frames(
-                    input_buffer=input_buffer,
-                    num_frames=next_read_size,
-                    read_upto_frame=end_idx,
-                )
-                if res is False:
-                    queue.put({
-                        "type": "END_PARTITION",
-                    })
-                    raise JobCancelledError("timeout while handling task")
-                if res is True:
-                    queue.put({
-                        "type": "END_PARTITION",
-                    })
-                    raise JobCancelledError("expected more data, didn't get any")
-                frames_read += res.num_frames
-                queue.put({
-                    "type": "FRAME",
-                    "start_idx": res.start_idx,
-                    "end_idx": res.end_idx,
-                    "first_frame_header": res.first_frame_header,
-                }, payload=np.frombuffer(res.buffer, dtype=np.uint8))
-            # FIXME: END_PARTITION in finally block?
-            queue.put({
-                "type": "END_PARTITION",
-            })
-
-    def start(self):
-        pass
-
-    def done(self):
-        pass
+class MerlinCommHandler(GenericCommHandler):
+    def get_conn_impl(self):
+        return self._conn.get_conn_impl()
 
 
 class MerlinLivePartition(Partition):
@@ -364,52 +263,19 @@ class MerlinLivePartition(Partition):
         # assert len(tiling_scheme) == 1
         tiling_scheme = tiling_scheme.adjust_for_partition(self)
         logger.debug("reading up to frame idx %d for this partition", self._end_idx)
+        to_read = self._end_idx - self._start_idx
 
-        queue = self._worker_context.get_worker_queue()
-        frames = get_frames_from_queue(
-            queue,
-            tiling_scheme,
-            self.shape.sig.to_tuple(),
-            dtype=dest_dtype
-        )
-
-        # special case: copy from the decode buffer into a larger partition buffer
-        # can be further optimized if needed (by directly decoding into said buffer)
-        if tiling_scheme.intent == "partition":
-            frame_stack = _accum_partition(
-                frames,
-                tiling_scheme,
-                self.shape.sig.to_tuple(),
-                dtype=dest_dtype
-            )
-            tile_shape = Shape(
-                frame_stack.shape,
-                sig_dims=2
-            )
-            tile_slice = Slice(
-                origin=(self._start_idx,) + (0, 0),
-                shape=tile_shape,
-            )
-            yield DataTile(
-                frame_stack,
-                tile_slice=tile_slice,
-                scheme_idx=0,
-            )
-            return
-
-        for frame_stack, start_idx in frames:
-            tile_shape = Shape(
-                frame_stack.shape,
-                sig_dims=2
-            )
-            tile_slice = Slice(
-                origin=(start_idx,) + (0, 0),
-                shape=tile_shape,
-            )
-            yield DataTile(
-                frame_stack,
-                tile_slice=tile_slice,
-                scheme_idx=0,
+        with QdGetFrames(
+            request_queue=self._worker_context.get_worker_queue(),
+            dtype=dest_dtype,
+            sig_shape=tuple(tiling_scheme[0].shape),
+        ) as frames:
+            yield from frames.get_tiles(
+                to_read=to_read,
+                start_idx=self._start_idx,
+                tiling_scheme=tiling_scheme,
+                roi=roi,
+                array_backend=array_backend,
             )
 
     def get_tiles(self, tiling_scheme, dest_dtype="float32", roi=None,
